arm: mx6_c1: Enable console use without serial
For now this enables the console on the hdmi output and probes
for usb keyboard for input on init.  We move the display setup
to early init so there is time for the user to see what they
should do.

This is a testing configuration while we are in the development
phase. The boot procedure should probably be expanded to catch
a hotkey to break out of the splash screen instead of showing
the console by default.

FIXME: USB Keyboard input can be laggy and inconsistent