board: cubox-i: Add FastBoot support to SPL
This adds three different boot modes to the SPL driver.  We store
the information in a general persistent 32-bit register, so it
will survive POR. The three modes are

RECOVERY BOOT: The default mode that loads the second stage u-boot
loader load a full set of supported drivers and do a normal bootloader
process loading/running boot.scr's, or uEnv.txt.
FAST BOOT: This mode attempts to load the kernel and dtb and jump
directly to them.  Setting up minimal devices by the bootloader.
DEVELOPER BOOT:  The same as recovery as far as functionality.  The
main difference being in how the workflow could work.

FASTBOOT works in this mode.  Userspace or Kernel needs to initiate
it by setting the register appropriately.  SPL sees this and
attempts to directly boot the Linux kernel.  Before it actually
boots the kernel it puts the register back to Recovery mode.  Therefore
if the kernel doesn't boot correctly and reset fastboot the user
won't be stuck in a boot loop.

DEVELOPER BOOT exists as a flag for developers to use to say,
"Never fastboot this kernel"  Mostly useful for tftbooting
kernels while doing development work and saving a few seconds on
boot isn't a big deal.