arm: mx6: PFD locking errata
The original fix for this errata sometime actually caused the
problem, gating and ungating the clocks caused instability in
on of the PFD's.  After further debugging we believe we have
narrowed down the circumstances that we need to cycle the pfd's
to the case where the state of the pfd's don't match.