imx6: spl: add alternative boot sector for RAW image access
Upstream has adopted a different offset than our image originally
used.  In order to migrate more easily to upstream uboot in the
future we will support both.  This should allow us to make the
change but still support legacy images.

We check the new sector offset first because it is farther into the
disk and avoids the situation where an old image with a valid
magic number sits at the old offset but is incomplete because a new
image has been written over it starting at the new offset.